refactor(types): promote trajectory data types to rllm.types
Move Step, Trajectory, Episode (training-extended), Action,
TrajectoryGroup, and _DEFAULT_TRAJ_NAME from rllm/agents/agent.py
into rllm/types.py as the single canonical home. The lightweight
Step/Trajectory/Episode previously in rllm/types.py were a strict
subset of the training subclasses, so the merge is purely additive
(extra fields default-empty in eval-only paths).

rllm/agents/agent.py is now a backward-compat re-export shim that
keeps only the BaseAgent ABC locally. All `from rllm.agents.agent
import …` callers continue to work unchanged via class identity:
`rllm.agents.agent.Episode is rllm.types.Episode`.

This is the first step of the AgentExecutionEngine + Agent +
Environment deprecation. Lazy imports of rllm.tools.tool_base,
rllm.engine.rollout, and rllm.workflows.workflow are kept inside
method bodies so rllm.types stays free of top-level legacy coupling.
